mina_hasher

Module roinput

Random oracle input structures and algorithms

Definition of random oracle input structure and methods for serializing into bytes and field elements

Structs

ROInput

Random oracle input structure

Constants

SER_HEADER_SIZE 🔒

Total number of bytes for the header of the serialized ROInput

SINGLE_HEADER_SIZE 🔒

Number of bytes for each part of the header of the serialized ROInput